Update 7
I managed to get majority of the Sprint Challenge completed, I just couldn't figure out how to show the number of residents in each LocationCard, number of episodes in the CharacterCard as well. Also not sure what is happening with the cards in Location and Episodes, I have been trying to make them wrap around so that they can be next to each other, as shown in the example.